First draft of nested components for rendering the group of summary cards for an array of Tables
